pyproject.toml
[tool.poetry]
name = "apex"
version = "0.1.0"
description = "A SaaS platform for building, training, and deploying AI chat agents"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "apex", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.11"
conduit-py = {path = "../conduit", develop = true}
sentence-transformers = ">=2.3.0"
pydantic = {extras = ["email"], version = "2.10.0"}
email-validator = "^2.0.0"
redis = "5.0.1"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.0.0"
python-multipart = "^0.0.9"
asyncpg = "^0.30.0"
pydantic-settings = "^2.0.0"
